+++ packages/agtree/src/parser/index.ts
@@ -1,5 +1,5 @@
 export { RuleParser } from './rule-parser';
-export { defaultParserOptions, type ParserOptions } from './options';
+export { defaultParserOptions, type ParserOptions, type OnParseError } from './options';
 export { CosmeticRuleParser } from './cosmetic/cosmetic-rule-parser';
 export { DomainListParser } from './misc/domain-list-parser';
 export { NetworkRuleParser } from './network/network-rule-parser';