+++ packages/examples/adguard-api-mv3/constants.ts
@@ -1,0 +1,1 @@
+export const ENABLED_FILTERS_IDS = ['2', '3', '4'];